Arena serialization
Added a Save Arena / Load Arena button to save the arena in the scene
and load an arena from file. If you want to test it, just open the game,
press Save and then press Load and check your Console for the location
of the save file.

Added files to control the behaviour of Balls, Spawn Points and Goals.
For now they don't have much except for PlayerID info but spawns are
already prepared to spawn balls (respawn them would simply be to change
the ball transform back to the spawn point's position).

C-Sharped some files a bit according to the usual Microsoft naming
convention (public variables are capitalized, etc.). Added some
comments.