Added skeleton of client/server networking
Created a simple UI to select whether we are server or client and start
server / connect to a server;

Once server starts, the server will wait for a client connection.

Once client is connected to server, the Open Space arena should be
loaded on both sides and player input should be restricted to only one
set of balls (different on each side).

After this we have to implement the actual game. I can't test this on
corporate network but it should be working (hopefully).