Basic networking
Networking now works fine between server and client. Both sides instance
the arena and are able to each create impulses on different sets of
balls.

Timer time left is being synchronized between players because we can
cause our own timer to stop (by moving the window, for example) and
server is always authoritative (although the player can also force timer
to stop by making the window lose focus, for example - this must be
fixed in the future but it's not that serious at this point in testing).