Physics tweaks, lights, refactors, camera
Refactored a few methods to use generic FindBallByName instead of
duplicating the same code over and over.
Added Light objects to arena serialization.
Tweaked physics a little bit so that the game is a bit more fun (it's
still not the same as the Kivy version, though).
Changed default "North" camera to straight top-down view.